requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
apache-airflow==2.10.2
apache-airflow-providers-amazon==9.0.0
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.18.0
apache-airflow-providers-fab==1.4.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.1
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-postgres==5.13.1
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.7.0
argcomplete==3.5.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.47
botocore==1.35.47
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==43.0.3
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.9
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.3
fastapi-cli==0.0.5
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.5.0
fsspec==2024.10.0
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.28.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
minio==7.2.9
more-itertools==10.5.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
prison==0.2.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
PyAthena==3.9.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
python-nvd3==0.16.0
python-slugify==8.0.4
python3-saml==1.16.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redshift-connector==2.1.3
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.3
rich-argparse==1.5.2
rpds-py==0.20.0
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
scramp==1.4.5
setproctitle==1.3.3
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.0
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tornado==6.4.1
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
watchtower==3.3.1
wcwidth==0.2.13
websockets==13.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.2.1
xmlsec==1.3.14
yarl==1.16.0
zipp==3.20.2
